# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "baby-greens",
    "description": "Sprout and microgreen tracker",
    "license": {
      "name": ""
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/push/key": {
      "get": {
        "tags": [
          "push"
        ],
        "operationId": "push_key",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushKey"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/push/subscriptions": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Save this device's push subscription, so reminders reach it.",
        "operationId": "subscribe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Saved"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "push"
        ],
        "summary": "Stop sending reminders to this device.",
        "operationId": "unsubscribe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnsubscribeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Removed, or it wasn't saved"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/push/test": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Send a test notification to all of your devices.",
        "operationId": "send_test",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResult"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Push notifications are off on this server (`push_disabled`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List open tasks on your active batches, soonest first.",
        "operationId": "list_tasks",
        "parameters": [
          {
            "name": "due_before",
            "in": "query",
            "description": "Only return tasks due at or before this time, like the end of the user's day.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskView"
                  }
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/tasks/{id}/complete": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Mark a task done. Finishing a care task schedules the next one. Finishing an advance task\nmoves the batch into its next step, and moving into harvest ends the batch.",
        "operationId": "complete_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Done"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The task is already done (`task_done`), or its batch has moved on (`batch_not_active`, `stale_task`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/tasks/{id}/snooze": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Push a task back. You get a new reminder when the snooze ends.",
        "operationId": "snooze_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnoozeTask"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Snoozed"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The task is done or its batch has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The time is in the past or more than a week away",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CareAction": {
        "type": "string",
        "enum": [
          "rinse",
          "water",
          "mist"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A stable code, like `not_found`, `invalid`, or `task_done`."
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "problems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Problem"
            },
            "description": "For `invalid`, what's wrong with each field."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "One thing wrong with a plant definition. `path` points at the field, like `steps[1].care[0]`.",
        "required": [
          "path",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "PushKey": {
        "type": "object",
        "properties": {
          "public_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "The VAPID public key, as base64url, for `pushManager.subscribe`. Null when push\nnotifications are off on this server."
          }
        }
      },
...
      "SnoozeTask": {
        "type": "object",
        "required": [
          "until"
        ],
        "properties": {
          "until": {
            "type": "string",
            "format": "date-time",
            "description": "When to remind you again. Within the next week."
          }
        },
        "additionalProperties": false
      },
...
      "StepAction": {
        "type": "string",
        "enum": [
          "soak",
          "sprout",
          "blackout",
          "light",
          "harvest"
        ]
      },
...
      "SubscriptionBody": {
        "type": "object",
        "description": "The JSON a browser's `PushSubscription.toJSON()` gives.",
        "required": [
          "endpoint",
          "keys"
        ],
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "keys": {
            "$ref": "#/components/schemas/SubscriptionKeys"
          }
        }
      },
      "SubscriptionKeys": {
        "type": "object",
        "required": [
          "p256dh",
          "auth"
        ],
        "properties": {
          "auth": {
            "type": "string"
          },
          "p256dh": {
            "type": "string"
          }
        }
      },
      "TaskAction": {
        "oneOf": [
          {
            "type": "object",
            "description": "Move the batch into the next step, whose action this is.",
            "required": [
              "action",
              "kind"
            ],
            "properties": {
              "action": {
                "$ref": "#/components/schemas/StepAction",
                "description": "Move the batch into the next step, whose action this is."
              },
              "kind": {
                "type": "string",
                "enum": [
                  "advance"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "action",
              "kind"
            ],
            "properties": {
              "action": {
                "$ref": "#/components/schemas/CareAction"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "care"
                ]
              }
            }
          }
        ],
        "description": "Serializes as `{ \"kind\": \"advance\", \"action\": \"light\" }` or `{ \"kind\": \"care\", \"action\": \"rinse\" }`."
      },
      "TaskView": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TaskAction"
          },
          {
            "type": "object",
            "required": [
              "id",
              "batch_id",
              "step_index",
              "due_at",
              "container",
              "plant_name"
            ],
            "properties": {
              "batch_id": {
                "type": "string",
                "format": "uuid"
              },
              "container": {
                "type": "string"
              },
              "due_at": {
                "type": "string",
                "format": "date-time"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "overdue_at": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "plant_name": {
                "type": "string"
              },
              "snoozed_until": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time",
                "description": "Set while the task is snoozed. `due_at` already counts the snooze."
              },
              "step_index": {
                "type": "integer",
                "minimum": 0
              }
            }
          }
        ]
      },
      "TestResult": {
        "type": "object",
        "required": [
          "sent",
          "removed"
        ],
        "properties": {
          "removed": {
            "type": "integer",
            "description": "Subscriptions removed because the push service no longer knows them.",
            "minimum": 0
          },
          "sent": {
            "type": "integer",
            "description": "Devices the push service accepted the message for.",
            "minimum": 0
          }
        }
      },
      "UnsubscribeBody": {
        "type": "object",
        "required": [
          "endpoint"
        ],
        "properties": {
          "endpoint": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bg_session"
      }
    }
  },
...
}